02Six_models_for_prediction_trained/requirements.txt
absl-py==1.4.0
accelerate==1.0.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.5.2
asttokens==3.0.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==24.3.0
backcall==0.2.0
beartype==0.20.2
beautifulsoup4==4.13.4
bleach==6.2.0
Brotli #@ file:///C:/b/abs_c415aux9ra/croot/brotli-split_1736182803933/work
cachetools==4.2.4
catalogue==2.0.10
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer #@ file:///croot/charset-normalizer_1721748349566/work
click==8.1.7
colorama==0.4.6
confection==0.1.5
contourpy #@ file:///C:/b/abs_f2u2o_8s9g/croot/contourpy_1732540071787/work
cycler #@ file:///tmp/build/80754af9/cycler_1637851556182/work
datasets==2.15.0
decorator==5.1.1
deepdiff==8.4.2
defusedxml==0.7.1
dill==0.3.7
distlib==0.3.9
distro==1.9.0
dnspython==2.6.1
einops==0.8.0
exceptiongroup==1.2.2
executing==2.1.0
fastcore==1.5.38
fastjsonschema==2.21.1
fasttransform==0.0.2
filelock==3.16.1
flatbuffers==24.3.25
fonttools #@ file:///C:/b/abs_4crkswws2h/croot/fonttools_1737040078745/work
frozenlist==1.6.0
fsspec==2023.10.0
ftfy==6.2.3
future==1.0.0
fuzzywuzzy==0.18.0
gast==0.3.3
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.66.0
graphviz==0.20.3
grpcio==1.68.1
h11==0.14.0
h5py==3.6.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.26.2
icecream==2.1.4
identify==2.6.10
idna==3.10
imageio==2.37.0
imbalanced-learn==0.11.0
imgaug==0.4.0
importlib_metadata==8.5.0
importlib_resources #@ file:///C:/b/abs_e9f8_0_2dq/croot/importlib_resources-suite_1720641112288/work
inflect==7.4.0
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.3
jiter==0.7.1
joblib==1.4.2
jsonlines==4.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keras==2.8.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keras-tuner==1.1.0
kiwisolver #@ file:///C:/b/abs_faf90xet7a/croot/kiwisolver_1737040915779/work
kt-legacy==1.0.5
lazy_loader==0.4
Levenshtein==0.27.1
libclang==18.1.1
loguru==0.7.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mergedeep==1.3.4
mistune==3.1.3
mkl-service==2.4.0
mkl_fft #@ file:///C:/Users/dev-admin/mkl/mkl_fft_1730823082242/work
mkl_random #@ file:///C:/Users/dev-admin/mkl/mkl_random_1730822522280/work
more-itertools==10.5.0
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.15
narwhals==1.38.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
networkx==3.0
nltk==3.9.1
nodeenv==1.9.1
numpy==1.25.2
oauthlib==3.2.2
openai==1.55.0
opencv-python==4.10.0.84
opt-einsum==3.3.0
orderly-set==5.4.0
packaging #@ file:///C:/b/abs_3by6s2fa66/croot/packaging_1734472138782/work
pandas==1.5.3
pandocfilters==1.5.1
parso==0.8.4
permetrics==1.3.0
pickleshare==0.7.5
pillow==11.2.1
platformdirs #@ file:///C:/b/abs_ddh15014or/croot/platformdirs_1744273060660/work
plum-dispatch==2.5.7
ply==3.11
pooch #@ file:///C:/b/abs_7acoiuwpey/croot/pooch_1731007069545/work
pre_commit==4.2.0
prettytable==3.11.0
promise==2.3
prompt_toolkit==3.0.48
propcache==0.3.1
protobuf==3.20.3
psutil==7.0.0
pure_eval==0.2.3
pyarrow==20.0.0
pyarrow-hotfix==0.7
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.0
pydantic_core==2.27.0
pydot==3.0.3
Pygments==2.18.0
pymongo==4.10.1
PyMuPDF==1.25.5
pyparsing #@ file:///C:/b/abs_40z8gyj9wi/croot/pyparsing_1731445739241/work
PyQt5_sip #@ file:///C:/b/abs_d1i1jr9fgt/croot/pyqt-split_1736540551157/work/pyqt_sip
PyQt6==6.7.1
PyQt6_sip #@ file:///C:/b/abs_f47pmp6xmn/croot/pyqt-split_1744804513582/work/pyqt_sip
PySocks #@ file:///C:/ci_310/pysocks_1642089375450/work
python-dateutil #@ file:///C:/b/abs_3au_koqnbs/croot/python-dateutil_1716495777160/work
python-Levenshtein==0.27.1
pytz==2025.2
PyWavelets==1.8.0
pywin32==310
PyYAML==6.0.2
pyzmq==26.4.0
RapidFuzz==3.13.0
referencing==0.36.2
regex==2024.11.6
requests #@ file:///C:/b/abs_c3508vg8ez/croot/requests_1731000584867/work
requests-oauthlib==2.0.0
rich==14.0.0
rpds-py==0.24.0
rsa==4.9
safetensors==0.5.3
scikit-image==0.19.3
scikit-learn==1.2.2
scipy==1.11.4
selectivesearch==0.4
setuptools-scm==8.1.0
shapely==2.0.6
shellingham==1.5.4
sip #@ file:///C:/b/abs_5cto136kse/croot/sip_1738856220313/work
six==1.15.0
sniffio==1.3.1
soupsieve==2.7
srsly==2.5.1
stack-data==0.6.3
sympy==1.13.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.2.2
tensorflow-datasets==3.0.0
tensorflow-estimator==2.10.0
tensorflow-gpu==2.8.0
tensorflow-gpu-estimator==2.3.0
tensorflow-hub==0.3.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-metadata==1.14.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.5.0
tifffile==2025.3.30
timm==1.0.11
tinycss2==1.4.0
tokenizers==0.15.2
tomli #@ file:///C:/Windows/TEMP/abs_ac109f85-a7b3-4b4d-bcfd-52622eceddf0hy332ojo/croots/recipe/tomli_1657175513137/work
torch==2.1.2+cu118
torch-snippets==0.450
torchaudio==2.1.2+cu118
torchinfo==1.8.0
torchvision==0.16.2+cu118
tornado #@ file:///C:/b/abs_7cyu943ybx/croot/tornado_1733960510898/work
tqdm==4.67.0
traitlets==5.14.3
transformers==4.35.2
typeguard==4.4.0
typer==0.15.3
typing==3.7.4.3
typing_extensions==4.12.2
unicodedata2 #@ file:///C:/b/abs_dfnftvxi4k/croot/unicodedata2_1736543771112/work
urllib3==2.2.3
virtualenv==20.31.1
wasabi==1.1.3
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.6
win-inet-pton #@ file:///C:/ci_310/win_inet_pton_1642658466512/work
win32_setctime==1.2.0
wrapt==1.12.1
xmltodict==0.14.2
xxhash==3.5.0
yacs==0.1.8
yarl==1.20.0
zipp #@ file:///C:/b/abs_2fcpcg89my/croot/zipp_1732630746332/work
